api/requirements.txt
fastapi
dnspython
pydantic
email-validator